dashboard/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.7
redis==4.6.0
plotly==5.16.1
pandas==2.0.3
openpyxl==3.1.2